Move design system elements inside InstUI namespace. Fix code review findings.
