--- Core/Core/DesignSystem/BaseScreen.swift
@@ -1,257 +1,0 @@
-//
-// This file is part of Canvas.
-// Copyright (C) 2024-present  Instructure, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as
-// published by the Free Software Foundation, either version 3 of the
-// License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-//
-
-import SwiftUI
-
-public enum ScreenState: String, Equatable, Identifiable, CaseIterable {
-    case loading
-    case error
-    case empty
-    case data
-
-    public var id: String { rawValue }
-}
-
-/**
- This structure encapsulates properties that alter the behavior of the screen.
- Properties were collected in a mind that those won't change and are static to a specific screen.
- */
-public struct BaseScreenConfig: Equatable {
-    public let refreshable: Bool
-    public let showsIndicators: Bool
-    public let scrollAxes: Axis.Set
-    public let errorPandaConfig: InteractivePanda.Config
-    public let emptyPandaConfig: InteractivePanda.Config
-
-    /**
-     - parameters:
-         - refreshable: Controls whether pull-to-refresh is available in the data state. Error and empty states are always refreshable.
-         - showsIndicators: Whether to show scroll indicators in the data state.
-         - scrollAxes: The scrollable axes in the data state.
-     */
-    public init(
-        refreshable: Bool = true,
-        showsIndicators: Bool = true,
-        scrollAxes: Axis.Set = .vertical,
-        errorPandaConfig: InteractivePanda.Config = .error(),
-        emptyPandaConfig: InteractivePanda.Config = .empty()
-    ) {
-        self.refreshable = refreshable
-        self.showsIndicators = showsIndicators
-        self.scrollAxes = scrollAxes
-        self.errorPandaConfig = errorPandaConfig
-        self.emptyPandaConfig = emptyPandaConfig
-    }
-}
-
-public struct BaseScreen<Content>: View where Content: View {
-    public typealias RefreshCompletion = () -> Void
-    public static var EmptyRefreshAction: (RefreshCompletion) -> Void {
-        { $0() }
-    }
-
-    @Environment(\.sizeCategory) private var sizeCategory
-    private let state: ScreenState
-    private let config: BaseScreenConfig
-    private let refreshAction: (@escaping RefreshCompletion) -> Void
-    private let content: (GeometryProxy) -> Content
-
-    /**
-     - parameters:
-         - refreshAction: A block that gets called when the user performs a pull-to-refresh action. The parameter of this block is a completion callback that needs to be called one time on the main thread to finish the refresh animation.
-         - content: The view to be rendered in the data state. The block receives a GeometryProxy for convenience that fills the available space. The content is embedded into a scroll view and has a .backgroundLightest background color.
-     */
-    public init(
-        state: ScreenState,
-        config: BaseScreenConfig = .init(),
-        refreshAction: @escaping (@escaping RefreshCompletion) -> Void = EmptyRefreshAction,
-        content: @escaping (GeometryProxy) -> Content
-    ) {
-        self.state = state
-        self.config = config
-        self.refreshAction = refreshAction
-        self.content = content
-    }
-
-    public var body: some View {
-        ZStack {
-            switch state {
-            case .loading:
-                loadingIndicator
-            case .error:
-                panda(config: config.errorPandaConfig)
-            case .empty:
-                panda(config: config.emptyPandaConfig)
-            case .data:
-                GeometryReader { geometry in
-                    let content = content(geometry)
-
-                    if config.refreshable {
-                        RefreshableScrollView(
-                            config.scrollAxes,
-                            showsIndicators: config.showsIndicators,
-                            content: {
-                                content
-                            },
-                            refreshAction: refreshAction
-                        )
-                        .frame(width: geometry.size.width, height: geometry.size.height)
-                    } else {
-                        ScrollView(
-                            config.scrollAxes,
-                            showsIndicators: config.showsIndicators
-                        ) {
-                            content
-                        }
-                        .frame(width: geometry.size.width, height: geometry.size.height)
-                    }
-                }
-                .background(Color.backgroundLightest)
-            }
-        }
-        .animation(.default, value: state)
-    }
-
-    private var loadingIndicator: some View {
-        ProgressView()
-            .progressViewStyle(.indeterminateCircle(color: Color(Brand.shared.primary)))
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .background(Color.backgroundLightest)
-    }
-
-    private func panda(config: InteractivePanda.Config)
-    -> some View {
-        GeometryReader { geometry in
-            RefreshableScrollView(
-                content: {
-                    InteractivePanda(config: config)
-                    .frame(
-                        minWidth: geometry.size.width,
-                        minHeight: geometry.size.height,
-                        alignment: .center
-                    )
-                },
-                refreshAction: refreshAction)
-            .background(Color.backgroundLightest)
-        }
-    }
-}
-
-#if DEBUG
-
-#Preview("Loading") {
-    BaseScreen(state: .loading) { _ in
-        Text(verbatim: "Data")
-    }
-}
-
-#Preview("Error") {
-    BaseScreen(state: .error) { _ in
-        Text(verbatim: "Data")
-    }
-}
-
-#Preview("Empty") {
-    BaseScreen(state: .empty) { _ in
-        Text(verbatim: "Data")
-    }
-}
-
-#Preview("Vertical Scrollable Data") {
-    BaseScreen(state: .data) { _ in
-        VStack(alignment: .leading) {
-            ForEach(0..<100) { index in
-                HStack {
-                    Text(verbatim: "Data \(index)")
-                    Spacer()
-                    Text(verbatim: "Action \(index)")
-                }
-            }
-        }
-        .padding()
-    }
-}
-
-#Preview("Horizontal Scrollable Data") {
-    BaseScreen(state: .data, config: .init(scrollAxes: .horizontal)) { _ in
-        HStack {
-            ForEach(0..<20) { index in
-                VStack {
-                    Text(verbatim: "Data \(index)")
-                    Spacer()
-                    Text(verbatim: "Action \(index)")
-                }
-            }
-        }
-        .padding()
-    }
-}
-
-#Preview("Vertical Small Data") {
-    BaseScreen(state: .data) { _ in
-        VStack(alignment: .leading) {
-            Text(verbatim: "Data")
-        }
-        .padding()
-    }
-}
-
-#Preview("Vertical Small Data Not Refreshable") {
-    BaseScreen(state: .data, config: .init(refreshable: false)) { _ in
-        VStack(alignment: .leading) {
-            Text(verbatim: "Non Refreshable Data")
-        }
-        .padding()
-    }
-}
-
-#Preview("Geometry Reader") {
-    BaseScreen(state: .data) { geometry in
-        VStack(spacing: 0) {
-            Color.red.frame(height: geometry.size.height / 3)
-            Color.white.frame(height: geometry.size.height / 3)
-            Color.green.frame(height: geometry.size.height / 3)
-        }
-    }
-}
-
-class ChangingViewModel: ObservableObject {
-    @Published var state = ScreenState.loading
-    private lazy var timer: Timer = Timer.scheduledTimer(withTimeInterval: 1.0, repeats: true) { [weak self] _ in
-        self?.state = ScreenState.allCases.randomElement()!
-    }
-
-    init() {
-        _ = timer
-    }
-}
-struct ChangingView: View {
-    @StateObject var viewModel = ChangingViewModel()
-
-    var body: some View {
-        BaseScreen(state: viewModel.state) { _ in
-            Text(verbatim: "Data")
-        }
-    }
-}
-
-#Preview("Transitions") {
-    ChangingView()
-}
-
-#endif
